# app/internal/docs/swagger.yaml
basePath: /
definitions:
  models.MonthDate:
    properties:
      time.Time:
        type: string
    type: object
  models.Subscription:
    properties:
      end_date:
        allOf:
        - $ref: '#/definitions/models.MonthDate'
        description: Optional end date.
      id:
        description: Subscription identifier.
        type: integer
      price:
        description: Monthly price.
        minimum: 0
        type: integer
      service_name:
        description: Service name.
        type: string
      start_date:
        allOf:
        - $ref: '#/definitions/models.MonthDate'
        description: Start date (month-year).
      user_id:
        description: Associated user ID.
        type: string
    required:
    - service_name
    - start_date
    - user_id
    type: object
  models.SummaryRequest:
    properties:
      from:
        allOf:
        - $ref: '#/definitions/models.MonthDate'
        description: Start of the period.
      service_name:
        description: Optional service filter.
        type: string
      to:
        allOf:
        - $ref: '#/definitions/models.MonthDate'
        description: End of the period.
      user_id:
        description: Optional user filter.
        type: string
    required:
    - from
    - to
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API для управления подписками пользователей
  title: Subscriptions API
  version: "1.0"
paths:
  /subscriptions/:
    get:
      description: Возвращает список подписок с пагинацией
      parameters:
      - description: Количество элементов на странице (по умолчанию 10)
        in: query
        name: limit
        type: integer
      - description: Смещение (по умолчанию 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'data: список подписок, limit, offset'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список подписок
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Создает новую подписку пользователя
      parameters:
      - description: Данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/models.Subscription'
      produces:
      - application/json
      responses:
        "201":
          description: Успешное создание
          schema:
            $ref: '#/definitions/models.Subscription'
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать подписку
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Удаляет подписку по ID
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Удалено
        "400":
          description: Некорректный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить подписку
      tags:
      - subscriptions
    get:
      description: Возвращает данные подписки по ID
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Найдена
          schema:
            $ref: '#/definitions/models.Subscription'
        "400":
          description: Некорректный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Не найдена
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить подписку по ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Обновляет данные существующей подписки
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/models.Subscription'
      produces:
      - application/json
      responses:
        "200":
          description: Обновлено
          schema:
            $ref: '#/definitions/models.Subscription'
        "400":
          description: Некорректные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить подписку
      tags:
      - subscriptions
  /subscriptions/summary:
    post:
      consumes:
      - application/json
      description: Возвращает общую сумму подписок за указанный период с учетом фильтров
      parameters:
      - description: Параметры периода и фильтров
        in: body
        name: summary
        required: true
        schema:
          $ref: '#/definitions/models.SummaryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Сумма подписок
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить сумму подписок за период
      tags:
      - subscriptions
swagger: "2.0"